Move git issue creator out of reporter #14
