+++ src/file_scanner_gcloudignore.py
@@ -42,14 +42,9 @@ class GCloudIgnoreFileScanner(FileScanner):
         for r in GCloudIgnoreFileScanner.REQUIRED_ENTRIES:
             if r not in entries:
                 result.missings.append(r)
-                result.problem.append("{} is not ignored".format(r))
                 result.status = ScanResultStatus.ERROR
         for r, pattern in GCloudIgnoreFileScanner.CONDITIONAL_ENTRIES:
             if r not in entries and self._contain_file(filelist, pattern):
                 result.missings.append(r)
-                result.problem.append("{} is not ignored".format(r))
                 result.status = ScanResultStatus.ERROR
-
-        if result.status == ScanResultStatus.ERROR:
-            result.remedy.append("Add the corresponding ignore entry.")
         return result
